DAS-2292: Add configuration for dropped variables.
It turns out for two reasons. We don't have spacecraft_overpass_time_utc.
The first is that it's huge, it's a 3km grid so 11568 x 4872 entries, and those
entries are 24 character String variables that is  1,352,623,104 bytes of data for
that one variable.  The second is that the NetCDF-C lib is not compressing
string data even if it's constant length.